Unify cinematic brightness
Each call to setBrightnessCin() will now spawn a C binary in the rootfs (`/opt/bin/cinematic_brightness`) that will take care of the job for `ipd`. It was becoming harder and harder to maintain the same code in three separate places (GUI, rootfs and `ipd`), so I decided to centralize the whole thing.